+++ cmd/root.go
@@ -26,7 +26,7 @@ Explain an aspect of Senzing.
 // ----------------------------------------------------------------------------
 
 var ContextVariablesForMultiPlatform = []option.ContextVariable{
-	option.ErrorId,
+	option.MessageID,
 	option.TtyOnly,
 }
 
@@ -66,13 +66,13 @@ func RunE(_ *cobra.Command, _ []string) error {
 
 	// Choose and run explainer.
 
-	if len(viper.GetString(option.ErrorId.Arg)) > 0 {
-		anExplainer = &explainer.ExplainerError{
-			ErrorId: viper.GetString(option.ErrorId.Arg),
-			TtyOnly: viper.GetBool(option.TtyOnly.Arg),
+	if len(viper.GetString(option.MessageID.Arg)) > 0 {
+		anExplainer = &explainer.MessageExplainer{
+			MessageID: viper.GetString(option.MessageID.Arg),
+			TtyOnly:   viper.GetBool(option.TtyOnly.Arg),
 		}
 	} else {
-		anExplainer = &explainer.ExplainerNull{}
+		anExplainer = &explainer.NullExplainer{}
 	}
 	return anExplainer.Explain(ctx)
 }