+++ explainer/explainer_null_test.go
@@ -1,23 +1,19 @@
-package explainer_test
+package explainer
 
 import (
 	"context"
 	"testing"
 
-	"github.com/senzing-garage/explain/explainer"
 	"github.com/stretchr/testify/assert"
 )
 
 // ----------------------------------------------------------------------------
-// Test interface functions
+// Test public functions
 // ----------------------------------------------------------------------------
 
-func TestExplainerNull(test *testing.T) {
+func TestNullExplainer(test *testing.T) {
 	ctx := context.TODO()
-	testObject := &explainer.ExplainerError{
-		ErrorId: "senzing-60010000",
-		TtyOnly: true,
-	}
+	testObject := &NullExplainer{}
 	err := testObject.Explain(ctx)
-	assert.Nil(test, err)
+	assert.NoError(test, err)
 }