+++ explainer/helpers.go
@@ -3,5 +3,5 @@ package explainer
 import "fmt"
 
 func page(uriSegment string) string {
-	return fmt.Sprintf("hub.senzing.com/%s/errors", uriSegment)
+	return fmt.Sprintf("senzing-garage.github.io/%s/errors", uriSegment)
 }