+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.2] - 2024-06-12
+
+### Changed in 0.2.2
+
+- From `ErrorID` to `MessageID`
+- From `ExplainerError` to `ErrorExplainer`
+- From `ExplainerNull` to `NullExplainer`
+
 ## [0.2.1] - 2024-04-19
 
 ### Changed in 0.2.1